Soluciones

+++ app/Http/Controllers/Aplicacion/crudSoluciones.php
@@ -19,6 +19,7 @@ use App\Models\Soluciontipoinnova;
 use App\Http\Requests\Solucion\Store1Post;
 use App\Http\Requests\Solucion\Store2Post;
 use App\Http\Requests\Solucion\Store3Post;
+use App\Http\Requests\Solucion\UpdatePost;
 class crudSoluciones extends Controller
 {
     //
@@ -44,16 +45,30 @@ class crudSoluciones extends Controller
         return back()->with('error', 'Solución no fue creada');
        
     }
-    public function update(Request $request,Solucion $solucion){
+    public function update(UpdatePost $request,Solucion $solucion){
+        if(Auth::id() != $solucion->user_id && (!Auth::user()->hasRole('admin') && !Auth::user()->hasRole('superadmin'))){
+            return back()->with('error', 'No ingresaste esta solución.');
+        }
+        $validatedData = $request->validated();
+        $solucion->update($validatedData);
+        if(isset($validatedData['archivo'])){
+            $imageName = Archivos::storeImagen($solucion->id, $validatedData['archivo'], 'soluciones');
+            $solucion->archivo = $imageName;
+            $solucion->save();
+        }
+
         if($request->get('continue')){
             $request->session()->put('step', '2');
-            return redirect()->route('app.soluciones.edit', [$solucion->problema_id, $solucion->id])->with(['status' => 'Innovación fase 1', 'method' => 'PUT']);
+            return redirect()->route('app.soluciones.edit', [$solucion->problema_id, $solucion->id])->with(['status' => 'Solución fase 1 completada con éxito', 'method' => 'PUT']);
         } else {
-            return redirect()->route('app.escritorio')->with(['status' => 'Innovación fase 1']);
+            return back()->with(['status' => 'Solución fase 1 completada con éxito, no se olvide de completarla más tarde']);
         }
 
     }
     public function updateFase2(Store2Post $request,Solucion $solucion){
+        if(Auth::id() != $solucion->user_id && (!Auth::user()->hasRole('admin') && !Auth::user()->hasRole('superadmin'))){
+            return back()->with('error', 'No ingresaste esta solución.');
+        }
         $validatedData = $request->validated();
         $solucion->update($validatedData);
         Soluciontipoinnova::where('solucion_id',$solucion->id)->delete();
@@ -74,6 +89,9 @@ class crudSoluciones extends Controller
 
     }
     public function updateFase3(Store3Post $request,Solucion $solucion){
+        if(Auth::id() != $solucion->user_id && (!Auth::user()->hasRole('admin') && !Auth::user()->hasRole('superadmin'))){
+            return back()->with('error', 'No ingresaste esta solución.');
+        }
         $validatedData = $request->validated();
         $solucion->update($validatedData);
         if($request->get('continue')){
@@ -83,5 +101,16 @@ class crudSoluciones extends Controller
             return redirect()->route('app.escritorio')->with(['status' => 'Solución fase 3 completada con éxito, se ha registrado su solucón exitosamente']);
         }
 
+    }
+
+    public function destroy(Solucion $solucion) {
+        // dd($problema->id);
+        if(Auth::id() != $solucion->user_id && (!Auth::user()->hasRole('superadmin'))){
+            return back()->with('error', 'No ingresaste esta solución.');
+        }
+
+        $solucion->delete();
+        Soluciontipoinnova::where('solucion_id',$solucion->id)->delete();
+        return redirect()->route('home')->with('status', 'Solución eliminada con éxito.');
     }
 }
